+++ solidity2/contracts/throw/20/LotterieThrow20.sol
@@ -29,7 +29,7 @@ contract LotterieThrow20 is LotterieThrow721 {
 
   function deffered_constructor(
     bool waitValue,
-    uint16 nb721,
+    uint8 nb721,
     address _token,
     uint paramsId,
     uint paramsPhaseId,
